feat: page별 data components 분리
